+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+
 - Added env var `TTY_COMPATIBLE` to override auto-detection of TTY support (See console.rst for details). https://github.com/Textualize/rich/pull/3675
 
 ### Changed
@@ -15,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - An empty `NO_COLOR` env var is now considered disabled. https://github.com/Textualize/rich/pull/3675
 - An empty `FORCE_COLOR` env var is now considered disabled. https://github.com/Textualize/rich/pull/3675
 - Rich tracebacks will now render notes on Python 3.11 onwards (added with `Exception.add_note`) https://github.com/Textualize/rich/pull/3676
-
+- Indentation in exceptions won't be underlined https://github.com/Textualize/rich/pull/3678
 
 ## [13.9.4] - 2024-11-01